+++ fpy2/libraries/__init__.py
@@ -5,4 +5,11 @@ FPy standard library.
 
 """
 
+###########################################################
+# Sublibraries
+
+# base library (auto-imported into the `fpy2` module)
+from . import base
+
+# core operations and utilities
 from . import core